WIP: DtbLoader: Convert to Application
This converts DtbLoader into an application that can chain-load the real
next stage.  It will look for "grubaa64.efi" in the same directory.

For example, if Boot#### variable points to \EFI\fedora\DtbLoader.efi
then this will look for \EFI\fedora\grubaa64.efi as the next stage.

Note that this also drops deferred handling of SMBIOS tables which was
needed as an EFI driver, which can be loaded before the SMBIOS tables
are installed.